Refactor code
Made code more compact by combining different functions into a single
object.

Also removed the need for a separate function for adding functionality
to the equals button. Previously, there had been an error when trying
to include it along with the other event handlers, so I had separated
it into its own function, but the error may have lied elsewhere as
there was no issue putting the logic back with the rest of the event
handlers.