Add support for use of backspace button
Clicking the backspace button will delete the last character entered
into the calculator. As characters are deleted, buttons will be
enabled or disabled accordingly, just as they would be as the user
is entering their calculation. That is, deleting a decimal point
character will enable the decimal button for use again, and deleting
a character that leaves a decimal point as the last character in the
display will disable the equals button and the operator buttons.