Improve behavior of previous calculation message
The previous calculation message will now also show the result of the
calculation. This change was made because previously, the result would
only be displayed in the current calculation display message, so when
the user modified the result in any way, either by deleting it or by
entering a new calculation, the previous calculation message would no
longer be showing the result of that calculation and would instead
point to the current calculation as if it were the result, when that
is no longer the case. This is a small quality of life improvement.